Move functions into new file InterpolationTargetDetail.hpp
This commit only moves code between InterpolationTargetReceiveVars.hpp
and the new InterpolationTargetDetail.hpp.
